Improving LLM evaluators (#25)
This PR adds notebooks for the two LLM evaluators in attempts to improve
them (for transcript eval and summarization eval).

Some key themes:
- I'm asking for a "reasoning" output, so that all criterias marked as
`0` gets an explanation. I also find that this approach gave insights to
why the LLM is marking something as 0, and helped me tweak prompts
accordingly to achieve what I want.
- I've switched over to GPT-4o, which gave a substantial improvement
over GPT-3.5 in pointing to specific examples when reasoning, and
adherence to prompt instructions.